refactor: make EventSource an abstract class
The change is motivated by tests I made with GraalVM native image. It seems that an abstract class instead of trait makes it easier for Native Image
to compile an app where wire-signals is a dependency. I know it should be fixed on the Native Image but I also wanted to make this change anyway,
because EventSource is in practice an abstract class - it's a common superclass for both EventStream and Signal, not only a trait enriching them
with some functionality and specifying what they can do.

This also means removal of ForcedEventSource which was a bit clunky way of making sure that .disablePauseWithContext() is called on every subscription
of the given stream/signal. Not a very usable feature and sometimes, I guess, might even introduce a bug.